Refactor search and integrate into admin
Previously, the admin search was being handled by Administrate and using
`LIKE` queries to get the job done. It has been replaced with PgSearch's
framework for single-table postgres full-text searching. This should be
a major improvement over the SQL-compliant text queries in both speed
and reliability, and uses the same engine configuration as the frontend
search.