Revamp homepage to present slideshow of events
This revamps the homepage to show a slideshow of all current episodes
and upcoming episodes, as well as "featured" episodes & performances at
the bottom. In the future, we'll be adding blog articles to the
"featured" section as well as any other kind of content we deem
necessary.

The FullPage.js layout didn't work as well as I thought. Either way, I
still wanted this left-to-right slideshow up top for the current
episode(s) and future episodes. I think we could use fullPage when we
have more article content or possibly more recordings? It seems like we
just don't have enough content right now to warrant using it. But I'm
keeping it on my radar because it's a stellar plugin and I think it
would look amazing on our site someday.

I haven't added past episodes or anything, because I was unsure whether
we should just throw all the past ones up there or whether we should
just put the ones we select in the "Features" section below.

Resolves #61